% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.bias.R
\docType{data}
\name{binom.bias}
\alias{binom.bias}
\title{EBF bias for binomial tests}
\format{
A vector with the bias for each sample size from 1 to 100.
}
\usage{
binom.bias
}
\description{
Bias in the log posterior marginal likelihood for binomial tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{\link{compute.binom.bias}} function using enumeration.
Sample sizes range from 1 to 100.
The uniform prior is used.
}
\references{
Dudbridge F (2023) Empirical Bayes factors for common hypothesis tests.
arXiv:2301.11057
}
\author{
Frank Dudbridge
}
\keyword{datasets}
