% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.f.bias.R
\name{compute.f.bias}
\alias{compute.f.bias}
\title{Compute EBF bias for F tests}
\usage{
compute.f.bias(df1 = 1:100, df2 = 1:100)
}
\arguments{
\item{df1}{Vector of numerator degrees of freedom for which to compute bias.}

\item{df2}{Vector of denominator degrees of freedom for which to compute bias.}
}
\value{
A matrix with the biases corresponding to the elements of
\code{df1} and \code{df2}.
}
\description{
Computes the bias in the log posterior marginal likelihood for F tests.
}
\details{
Pre-computed results are in the \code{ebf::f.bias} object,
for degrees of freedom from 1 to 100 in either numerator or denominator.
The bias is computed by numerically integrating over a bivariate F distriubtion.
This requires the \code{cubature} library.
}
\references{
Dudbridge F (2023) Empirical Bayes factors for common hypothesis tests.
arXiv:2301.11057
}
\author{
Frank Dudbridge
}
