% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.nbinom.bias.R
\name{compute.nbinom.bias}
\alias{compute.nbinom.bias}
\title{Compute EBF bias for negative binomial tests}
\usage{
compute.nbinom.bias(
  size = 1:100,
  xmin = 0,
  xmax = 1,
  shape = 1,
  complement = FALSE,
  nsample = 1000,
  seed = 0
)
}
\arguments{
\item{size}{Vector of sample sizes for which to estimate bias.}

\item{xmin}{Minimum probability of interval hypothesis.}

\item{xmax}{Maximum probability of interval hypothesis.}

\item{shape}{Parameter of symmetrical Beta prior distribution.}

\item{complement}{If \code{TRUE}, computes the bias for the complement
of the interval hypothesis.}

\item{nsample}{Number of random samples taken to estimate bias.}

\item{seed}{Random number seed.}
}
\value{
A vector with the biases corresponding to the elements of
\code{size}.
}
\description{
Computes the bias in the log posterior marginal likelihood for negative binomial tests.
}
\details{
Pre-computed results are in the \code{ebf::nbinom.bias} object,
for sample sizes 1 to 100 with uniform prior.

Full enumeration of the sample space can be very time-consuming.
Therefore the bias is estimated by Monte Carlo
simulations of observed and replicate data from the negative binomial distribution.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
