% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.p.bias.R
\name{compute.p.bias}
\alias{compute.p.bias}
\title{Compute EBF bias for p-values}
\usage{
compute.p.bias(shape = 1, nsample = 0)
}
\arguments{
\item{shape}{Second parameter for Beta distribution of p-values.}

\item{nsample}{Number of random samples taken to estimate bias.
If 0 (default), analytic calculation is made.}
}
\value{
The bias for the given shape parameter.
This is about log(5/2) = 0.92 for the bulk of the distribution.
}
\description{
Computes the bias in the log posterior marginal likelihood for p-values.
}
\details{
The p-values are assumed to follow a Beta(1, \code{shape}) distribution,
where \code{shape}>=1.
The bias turns out to be about log(5/2) = 0.92 for most values of \code{shape}.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
