% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.poisson.bias.R
\name{compute.poisson.bias}
\alias{compute.poisson.bias}
\title{Compute EBF bias for Poisson tests}
\usage{
compute.poisson.bias()
}
\value{
The maximum bias over the rate parameter, which is 0.5003284.
}
\description{
Computes the bias in the log posterior marginal likelihood for Poisson tests.
}
\references{
Dudbridge F (submitted) A scale of evidence for likelihood ratios.

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
