% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.t.bias.R
\name{compute.t.bias}
\alias{compute.t.bias}
\title{Compute EBF bias for t tests}
\usage{
compute.t.bias(df = 1:100)
}
\arguments{
\item{df}{Vector of degrees of freedom for which to compute bias.}
}
\value{
A vector with the biases corresponding to \code{df}.
}
\description{
Computes the bias in the log posterior marginal likelihood for t tests.
}
\details{
Pre-computed results for 1 to 100 degrees of freedom are in the
\code{ebf::t.bias} object.
The bias is computed by numerically integrating over a bivariate t distriubtion.
This requires the \code{cubature} library.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in planning) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
