% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.binom.R
\name{ebf.binom}
\alias{ebf.binom}
\title{Empirical Bayes factors for binomial tests}
\usage{
ebf.binom(
  x,
  size,
  index = NULL,
  h0 = c(0.5, 0.5),
  h1 = NULL,
  shrink = FALSE,
  npoints = 1000,
  shape = 1
)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{size}{Vector containing the numbers of trials in each test.
The numbers of successes are contained in \code{x}.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated only for
those elements. If \code{NULL}, EBFs are calculated for all elements.}

\item{h0}{A vector with two elements, giving the lower and upper
bounds of the null hypothesis.}

\item{h1}{A vector with two elements, giving the lower and upper bounds of
the alternative hypothesis.  If \code{NULL} (default), the alternative
hypothesis is the complement of \code{h0}.

The default test has \code{h0=c(0.5,0.5)}, with two-sided alternative.
For one-sided alternatives, use \code{h1=c(0.5,1)} or \code{h1=c(0,0.5)}.
To test higher values against lower values, use \code{h0=c(0,0.5)}.
In this case \code{h1} defaults to \code{c(0.5,1)}.
To test lower values against higher values, use \code{h0=c(0.5,1)}.}

\item{shrink}{If \code{TRUE}, uses the entire distribution of \code{x}
to calculate each individual EBF.

If \code{shrink==TRUE} the calculation is of order \code{length(x)^2}
and may be very time-consuming.  To reduce the computation a subset of
\code{x} may be used to calculate the shrinkage EBFs.
By default a maximum of 1000 points is used, taken from the quantiles of the
simple EBFs.  This number can be varied with the \code{npoints} option.

Further reduction can be achieved by calculating EBFs only for the elements
of \code{x} given by \code{index}.  The full distribution of \code{x} is used
in calculating those EBFs.}

\item{npoints}{Number of points to use in calculate shrinkage EBFs.
Default is \code{min(length(x), 1000)}.}

\item{shape}{Parameter of the prior distribution Beta(\code{shape},\code{shape}).}
}
\value{
An object of class "ebf" containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence, that is
     on the log scale with base 3.73.  One unit updates
     weaker belief to stronger belief, or weaker disbelief to weaker belief,
     according to a rational definition of weaker and stronger belief.
     See references for further details.}
    \item{\code{ebf.shrink} Shrunken EBFs if specified by \code{shrink==TRUE}.}
    \item{\code{ebf.shrink.units} Expressed as units of evidence.}
    \item{\code{p} P-values.  This is empty if \code{h0} is an interval.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for binomial tests of proportions.
}
\details{
This provides an exact but conservative EBF.
If a normal approximation to the binomial is acceptable,
\code{\link{ebf.norm}} is faster.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in planning) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
