% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.f.R
\name{ebf.f}
\alias{ebf.f}
\title{Empirical Bayes factors for F tests}
\usage{
ebf.f(x, df1, df2, tails = 1, index = NULL, bias = NULL)
}
\arguments{
\item{x}{Vector of statistics.}

\item{df1}{Vector of numerator degrees of freedom.}

\item{df2}{Vector of denominator degrees of freedom.}

\item{tails}{Default is 1 for a one-tailed test as in ANOVA.
Can be 2 for a general variance ratio test.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated for
those elements only. If \code{NULL}, EBFs are calculated for all elements.}

\item{bias}{Vector of biases for each test.}
}
\value{
A data frame containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{p} P-values.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for F tests.
}
\details{
The EBF includes bias adjustments to the log posterior marginal likelihoods.
Pre-computed adjustments are used for \code{df1} and \code{df2} from 1 to 100.
For larger values, \code{\link{compute.f.bias}} is used to calculate an
accurate adjustment.  Pre-computed values can be supplied in the \code{bias} parameter.

EBFs are calculated in favour of \code{h1}.
}
\references{
Dudbridge F (2022) A scale of interpretation for likelihood ratios and Bayes factors.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors for common hypothesis tests.
}
\author{
Frank Dudbridge
}
