% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.poisson.R
\name{ebf.poisson}
\alias{ebf.poisson}
\title{Empirical Bayes factors for Poisson tests}
\usage{
ebf.poisson(
  x,
  interval = 1,
  index = NULL,
  h0 = 1,
  h1 = NULL,
  shrink = FALSE,
  npoints = 1000
)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{interval}{Vector containing the interval lengths in each test.
The numbers of events are contained in \code{x}.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated only for
those elements. If \code{NULL}, EBFs are calculated for all elements.}

\item{h0}{If a scalar, the value of a point null hypothesis.
If a vector with two elements, the lower and upper
bounds of the null hypothesis.}

\item{h1}{If a scalar, the value of a point alternative hypothesis.
If a vector with two elements, the lower and upper bounds of
the alternative hypothesis.  If \code{NULL} (default), the alternative
hypothesis is the complement of \code{h0}.

The default test has \code{h0=1}, with two-sided alternative.
For one-sided alternatives, use \code{h1=c(1,Inf)} or \code{h1=c(0,1)}.
To test higher values against lower values, use \code{h0=c(0,1)}.
In this case \code{h1} defaults to \code{c(1,Inf)}.
To test lower values against higher values, use \code{h0=c(1,Inf)}.}

\item{shrink}{If \code{TRUE}, uses all the data in \code{x}
to calculate each individual EBF.

The calculation is of order \code{length(x)^2}
and may be very time-consuming.  To reduce the computation a subset of
\code{x} may be used to calculate the shrinkage EBFs.
By default a maximum of 1000 points is used, taken from the quantiles of the
simple EBFs.  This number can be varied with the \code{npoints} parameter.

Further reduction can be achieved by calculating EBFs only for the elements
of \code{x} given by \code{index}.  All the points in \code{x} are used
in calculating those EBFs, unless varied by \code{npoints}.}

\item{npoints}{Number of points to use in calculate shrinkage EBFs.
Default is \code{min(length(x), 1000)}.}
}
\value{
A data frame containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence, that is
     on the log scale with base 3.73.  One unit updates
     weaker belief to stronger belief, or weaker disbelief to weaker belief,
     according to a rational definition of weaker and stronger belief.
     See references for further details.}
    \item{\code{p} P-values.  This is empty if \code{h0} is an interval.}
    \item{\code{p.log10} Expressed in -log10 units.}
    \item{\code{ebf.shrink} Shrunken EBFs if specified by \code{shrink==TRUE}.}
    \item{\code{ebf.shrink.units} Expressed as units of evidence.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for Poisson tests of rates.
}
\details{
The EBF includes bias adjustments to the log posterior marginal likelihoods.
The adjustment depends on the unknown rate parameter.
The maximum value is 0.5346, corresponding to a rate of 7.301.
This is used to give a conservative EBF.

If a normal approximation is acceptable, use \code{\link{ebf.norm}}.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in planning) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
