% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBF.R
\name{EBF}
\alias{EBF}
\title{EBF: A function to compute the Empirical Bayes Factor from model estimates}
\usage{
EBF(theta, sig, tau, logarithm = T)
}
\arguments{
\item{theta}{a matrix with the random effects estimates (name the columns, so the vector of EBFs will also be named)}

\item{sig}{a list containing the covariance matrix of the random effect, the list should have the same length as the number of random effects}

\item{tau}{a vector with the estimates of the variance of the random effects}

\item{logarithm}{a logical parameter that indicates whether to compute the EBF in the log scale or not}
}
\value{
a vector with the EBFs
}
\description{
EBF: A function to compute the Empirical Bayes Factor from model estimates
}
\examples{

library(EBF)
#Fitting the model
library(rstanarm)

data(example)

model <- stan_lmer(y ~ 1 + (1 | groupA) + (1 | groupB), data = example,
                   chains = 1)

posterior_samples_array <- as.data.frame(model)

b <- data.frame(groupA = colMeans(posterior_samples_array[,paste0("b[(Intercept) groupA:", 1:10, "]")]), #group A
                groupB = colMeans(posterior_samples_array[,paste0("b[(Intercept) groupB:", 1:10, "]")]) #group B
)

covb <- list(groupA = cov(posterior_samples_array[,paste0("b[(Intercept) groupA:", 1:10, "]")]),
             groupB = cov(posterior_samples_array[,paste0("b[(Intercept) groupB:", 1:10, "]")]))

sig <- colMeans(posterior_samples_array[,c("Sigma[groupA:(Intercept),(Intercept)]","Sigma[groupB:(Intercept),(Intercept)]")])

EBF(b, covb, sig)

}
\references{
Vieira, F., Zhao, H., & Mulder, J. (2024). To Vary or Not To Vary: A Simple Empirical Bayes Factor for Testing Variance Components. arXiv preprint arXiv:2410.14459.

Vieira, F., Leenders, R., McFarland, D., & Mulder, J. (2024). A Bayesian actor-oriented multilevel relational event model with hypothesis testing procedures. Behaviormetrika, 51(1), 37-74.
}
