% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.t.R
\name{ebf.t}
\alias{ebf.t}
\title{Empirical Bayes factors for t tests}
\usage{
ebf.t(
  x,
  se = NULL,
  h0 = c(0, 0),
  h1 = NULL,
  df = NULL,
  shrink = FALSE,
  index = NULL
)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{se}{Vector of standard errors associated with \code{x}.}

\item{h0}{A vector with two elements, giving the lower and upper
bounds of the null hypothesis.}

\item{h1}{A vector with two elements, giving the lower and upper bounds of
the alternative hypothesis to \code{h0}.  If unspecified, the alternative
hypothesis is the complement of \code{h0}.}

\item{df}{Vector of degrees of freedom.  Defaults to \code{Inf}.}

\item{shrink}{If \code{TRUE}, uses information from all elements of \code{x}
when calculating each individual EBF.}

\item{index}{Vector of indices selecting a subset of \code{x}.}
}
\value{
An object of class "ebf" containing the following components:
  \itemize{
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} \code{ebf} expressed as units of evidence.}
    \item{\code{ebf.shrink} Shrunken EBFs if specified by \code{shrink==TRUE}.}
    \item{\code{ebf.shrink.units} \code{ebf.shrink} expressed as units of evidence.}
    \item{\code{p} P-values.}
    \item{\code{p.log10} -log10 of \code{p}.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for univariate t tests.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.
}
\author{
Frank Dudbridge
}
