% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.bias.R
\docType{data}
\name{f.bias}
\alias{f.bias}
\title{EBF bias for F tests}
\format{
A matrix with the bias for each degree of freedom from 1 to 100
in either numerator or denominator.
}
\usage{
f.bias
}
\description{
Bias in the log posterior marginal likelihood for F tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{\link{compute.f.bias}} function using numerical integration.
Degrees of freedom range from 1 to 100.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
\keyword{datasets}
