% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinom.bias.R
\docType{data}
\name{nbinom.bias}
\alias{nbinom.bias}
\title{EBF bias for negative binomial tests}
\format{
A vector with the bias for each sample size from 1 to 100.
}
\usage{
nbinom.bias
}
\description{
Bias in the log posterior marginal likelihood for negative binomial tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{\link{compute.nbinom.bias}} function using Monte Carlo simulation.
Minimum sample sizes (ie fixed number of events) range from 1 to 100.
The uniform prior is used.
}
\references{
Dudbridge F (2023) Empirical Bayes factors for common hypothesis tests.
arXiv:2301.11057
}
\author{
Frank Dudbridge
}
\keyword{datasets}
