% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.EBF.R
\name{print.EBF}
\alias{print.EBF}
\title{print.EBF generic function to print the results of the EBF function}
\usage{
\method{print}{EBF}(x, ...)
}
\arguments{
\item{x}{an object of class EBF}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
print.EBF generic function to print the results of the EBF function
}
