% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.bias.R
\docType{data}
\name{t.bias}
\alias{t.bias}
\title{EBF bias for t tests}
\format{
A data frame with degrees of freedom in the first column
and estimated bias in the second .
}
\usage{
t.bias
}
\description{
Bias in the posterior mean likelihood for t tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{compute.t.bias} function using numerical integration.
Degrees of freedom range from 1 to 100.
}
\keyword{datasets}
