% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.bias.R
\docType{data}
\name{binom.bias}
\alias{binom.bias}
\title{EBF bias for binomial tests}
\format{
A data frame with the bias for each sample size from 1 to 100.
}
\usage{
binom.bias
}
\description{
Bias in the posterior marginal likelihood for binomial tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{compute.binom.bias} function using enumeration.
Sample sizes range from 1 to 100.
The uniform prior is used.
}
\details{
The bias depends on the true success probability.
As this is unknown, the maximum bias over success probabilities is taken
as a conservative estimate.
}
\keyword{datasets}
