% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.binom.bias.R
\name{compute.binom.bias}
\alias{compute.binom.bias}
\title{Compute EBF bias for binomial tests}
\usage{
compute.binom.bias(n = 1:100, shape = 1)
}
\arguments{
\item{n}{Vector of sample sizes for which to compute bias.}

\item{shape}{Parameter of symmetrical Beta prior distribution.}
}
\value{
A vector with the biases corresponding to the elements of
\code{n}.
}
\description{
Computes the bias in the log posterior marginal likelihood for binomial tests.
}
\details{
Pre-computed results are in the \code{ebf::binom.bias} object,
for sample sizes 1 to 100 with uniform prior.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
