% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.poisson.bias.R
\name{compute.poisson.bias}
\alias{compute.poisson.bias}
\title{Compute EBF bias for Poisson tests}
\usage{
compute.poisson.bias(
  xmin = 0,
  xmax = Inf,
  shape = 1,
  rate = 0,
  quant = 1e-04,
  limit = 100
)
}
\arguments{
\item{xmin}{Minimum probability of interval hypothesis.}

\item{xmax}{Maximum probability of interval hypothesis.}

\item{shape}{Shape parameter of prior Gamma distribution.}

\item{rate}{Rate parameter of prior Gamma distribution.}

\item{quant}{Minimum quantile of Poisson observations for which to compute bias.}

\item{limit}{Upper limit of observed data.}
}
\value{
The bias, which is 0.5 when \code{xmax==Inf} and 0 otherwise.
}
\description{
Computes the bias in the log posterior marginal likelihood for Poisson rates.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
