% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.chisq.R
\name{ebf.chisq}
\alias{ebf.chisq}
\title{Empirical Bayes factors for chi-squared tests}
\usage{
ebf.chisq(x, df = NULL, index = NULL)
}
\arguments{
\item{x}{Vector of chi-squared statistics for which EBFs will be calculated.}

\item{df}{Vector of degrees of freedom. Defaults to 1.}

\item{index}{Vector of indices selecting a subset of \code{x}.}
}
\value{
An object of class "ebf" containing the following components:
  \itemize{
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} \code{EBF} expressed as units of evidence.}
    \item{\code{p} P-values.}
    \item{\code{p.log10} -log10 of \code{p}.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for chi-squared tests.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.
}
\author{
Frank Dudbridge
}
