% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.chisq.R
\name{ebf.chisq}
\alias{ebf.chisq}
\title{Empirical Bayes factors for chi-squared tests}
\usage{
ebf.chisq(x, df = 1, index = NULL)
}
\arguments{
\item{x}{Vector of statistics.}

\item{df}{Vector of degrees of freedom.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated for
those elements only. If \code{NULL}, EBFs are calculated for all elements.}
}
\value{
A data frame containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{p} P-values.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for chi-squared tests.
}
\details{
The assumption is that a likelihood can be defined such that the chi-squared
is approximately twice the difference between the maximised
log-likelihood and the null log-likelihood.  Then it can be shown that
\code{EBF = exp((chi2 - df)/2) / (sqrt(2)^df)}.

EBFs are calculated in favour of \code{h1}.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
