% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.chisq.R
\name{ebf.chisq}
\alias{ebf.chisq}
\title{Empirical Bayes factors for chi-squared tests}
\usage{
ebf.chisq(x, df = 1, index = NULL)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{df}{Vector of degrees of freedom. Defaults to 1.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated only for
those elements. If \code{NULL}, EBFs are calculated for all elements.}
}
\value{
A data frame containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{p} P-values.  This is empty if \code{h0} is an interval.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for chi-squared tests.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in planning) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
