% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.f.R
\name{ebf.f}
\alias{ebf.f}
\title{Empirical Bayes factors for F tests}
\usage{
ebf.f(x, df1, df2, index = NULL, bias = NULL)
}
\arguments{
\item{x}{Vector of statistics.}

\item{df1}{Vector of numerator degrees of freedom.}

\item{df2}{Vector of denominator degrees of freedom.}

\item{index}{Vector of indices in \code{x}.  EBFs are calculated for
those elements only. If \code{NULL}, EBFs are calculated for all elements.}

\item{bias}{Vector of biases for each test.}
}
\value{
A data frame containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{p} P-values.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for F tests of variance ratios.
}
\details{
The EBF includes bias adjustments to the log posterior marginal likelihoods.
Pre-computed adjustments are used for \code{df1} and \code{df2} from 1 to 100.
If either are 1, an equivalent two-sided t test is used.
For larger values, the asymptotic adjustment of 0.5 is used.
This may however be inaccurate when there is a large difference between
\code{df1} and \code{df2}.
In that case, \code{\link{compute.f.bias}} can be used to calculate an
accurate adjustment.  This can then be supplied here in the \code{bias} parameter.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
