% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.norm.R
\name{ebf.norm}
\alias{ebf.norm}
\title{Empirical Bayes factors for normal tests}
\usage{
ebf.norm(
  x,
  se = NULL,
  h0 = c(0, 0),
  h1 = NULL,
  shrink = FALSE,
  npoints = 1000,
  index = NULL
)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{se}{Vector of standard errors associated with \code{x}.  Default is 1.}

\item{h0}{A vector with two elements, giving the lower and upper
bounds of the null hypothesis.}

\item{h1}{A vector with two elements, giving the lower and upper bounds of
the alternative hypothesis.  If unspecified, the alternative
hypothesis is the complement of \code{h0}.}

\item{shrink}{If \code{TRUE}, uses the entire distribution of \code{x}
to calculate each individual EBF.}

\item{npoints}{Number of points to use in calculate shrinkage EBFs.
Default is \code{min(length(x), 1000)}.}

\item{index}{Vector of indices of elements of \code{x} for which to calculate shrinkage EBFs.}
}
\value{
An object of class "ebf" containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{ebf.shrink} Shrunken EBFs if specified by \code{shrink==TRUE}.}
    \item{\code{ebf.shrink.units} Expressed as units of evidence.}
    \item{\code{p} P-values.  This is empty if \code{h0} is an interval.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for univariate normal tests.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (unpublished) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
