% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.R
\name{ebf}
\alias{ebf}
\title{Empirical Bayes factors}
\usage{
ebf(
  x,
  se = NULL,
  h0 = c(0, 0),
  h1 = NULL,
  test = c("normal", "t", "chisq"),
  df = NULL,
  shrink = FALSE,
  npoints = 1000,
  index = NULL
)
}
\arguments{
\item{x}{Vector of statistics for which EBFs will be calculated.}

\item{se}{Vector of standard errors associated with \code{x}.  Default is 1.}

\item{h0}{A vector with two elements, giving the lower and upper
bounds of the null hypothesis.}

\item{h1}{A vector with two elements, giving the lower and upper bounds of
the alternative hypothesis.  If unspecified, the alternative
hypothesis is the complement of \code{h0}.}

\item{test}{Type of EBF to compute.  Can be "normal" (default), "t" or "chisq".}

\item{df}{Vector of degrees of freedom for t or chi-squared tests.
Defaults to \code{Inf} for t test and 1 for chi-squared tests.
For the t test, values of \code{df} over 100 are set to \code{Inf}, equivalent to the Normal test.}

\item{shrink}{If \code{TRUE}, uses the entire distribution of \code{x}
to calculate each individual EBF.}

\item{npoints}{Number of points to use in calculate shrinkage EBFs.
Default is \code{min(length(x), 1000)}.}

\item{index}{Vector of indices of elements of \code{x} for which to calculate shrinkage EBFs.}
}
\value{
An object of class "ebf" containing the following components:
  \itemize{
  \item{\code{index} Indices of elements in \code{x} for which EBFs are calculated.}
    \item{\code{ebf} Empirical Bayes factors for the elements of \code{x}.}
    \item{\code{ebf.units} Expressed as units of evidence.}
    \item{\code{ebf.shrink} Shrunken EBFs if specified by \code{shrink==TRUE}.}
    \item{\code{ebf.shrink.units} Expressed as units of evidence.}
    \item{\code{p} P-values.  This is empty if \code{h0} is an interval.}
    \item{\code{p.log10} Expressed in -log10 units.}
  }
}
\description{
Calculates empirical Bayes factors (EBFs) for a set of statistics.
Normal, t and chi-squared tests are available.
Two-sided, one-sided and interval hypotheses can be tested.
}
\details{
This is the main function in the package.
By default, the statistics are normally distributed with known variances.
If \code{se} is unspecified then the variances are assumed to be 1.

The default test has mean=0 as the null hypothesis, with two-sided alternative.
For one-sided alternatives, use \code{h1=c(0,-Inf)} or \code{h1=c(-Inf,0)}.
To test positive values against negative values, use \code{h0=c(-Inf,0)}.
By default \code{h1} is the complement of \code{h0}, in this case \code{c(0,Inf)}.
To test negative values against positive values, use \code{h0=c(0,Inf)}.


For chi-squared tests, \code{h0}, \code{h1} and \code{shrink} are ignored,

If \code{shrink==TRUE} the calculation is of order \code{length(x)^2}
and may be very time-consuming.  To reduce the computation a subset of
\code{x} may be used to calculate the shrinkage EBFs.
By default a maximum of 1000 points is used, taken from the quantiles of the
simple EBFs.  This number can be varied with the \code{npoints} option.

It can be restricted to a subset of \code{x} by using \code{index}.
The shrunken EBFs are only calculated for the specified elements of \code{x},
however the full distribution of \code{x} is used in calculating those EBFs.

Units of evidence are EBFs on the log scale with base 3.73.  One unit updates
weaker belief to stronger belief, or weaker disbelief to weaker belief,
according to an objective definition of weaker and stronger belief.
See Dudbridge (submitted) for further details.
}
\references{
Dudbridge F (submitted) Units of evidence and expected Bayes factors for
objective reporting of statistical evidence.

Dudbridge F (unpublished) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
