% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.R
\docType{package}
\name{ebf}
\alias{ebf}
\title{ebf}
\description{
Package for calculating empirical Bayes factors (EBFs) for a set of statistics.
Normal, t and chi-squared tests are available.
Two-sided, one-sided and interval hypotheses can be tested.
}
\section{Main functions}{


\code{\link{ebf.norm}}

\code{\link{ebf.t}}

\code{\link{ebf.chisq}}
}

