% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebf.R
\docType{package}
\name{ebf}
\alias{ebf}
\title{ebf}
\description{
Package for calculating empirical Bayes factors (EBFs) for a set of statistics.
Normal, t, binomial, negative binomial, Poisson, F and chi-squared tests are available.
For non-parametric tests, direct conversion of P-values to EBFs is possible.
Two-sided, one-sided and interval hypotheses can be tested.
}
\section{Main functions}{


\code{\link{ebf.norm}}

\code{\link{ebf.t}}

\code{\link{ebf.binom}}

\code{\link{ebf.nbinom}}

\code{\link{ebf.poisson}}

\code{\link{ebf.f}}

\code{\link{ebf.chisq}}

\code{\link{ebf.p}}
}

\references{
Dudbridge F (2022) A scale of interpretation for likelihood ratios and Bayes factors.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors for common hypothesis tests.
}
\author{
Frank Dudbridge
}
