% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBF.R
\name{EBF}
\alias{EBF}
\title{EBF: A function to compute Empirical Bayes Factors.}
\usage{
EBF(theta, var, Sigma, logarithm = T)
}
\arguments{
\item{theta}{a vector of random effects estimates, or a matrix with multiple the random effects estimates (name the columns, so the vector of EBFs will also be named).}

\item{var}{the error (posterior) covariance matrix of the random effects, or a list of covariance matrices of multiple random effects.}

\item{Sigma}{an estimate of a variance of the random effects, or a vector with the estimates of the variance of the random effects.}

\item{logarithm}{a logical parameter that indicates whether to compute the EBF in the log scale or not}
}
\value{
a vector with the EBFs
}
\description{
A function to compute the Empirical Bayes Factor (Vieira et al., 2024a, 2024b) from the random effects estimates, the random effects error (posterior) covariance matrix, and the estimated random effects variance.
}
\examples{

library(EBF)
#Fitting the model
library(rstanarm)

data(example)

model <- stan_lmer(y ~ 1 + (1 | groupA) + (1 | groupB), data = example,
                   chains = 1)

posterior_samples_array <- as.data.frame(model)

b <- data.frame(groupA = colMeans(posterior_samples_array[,paste0("b[(Intercept) groupA:", 1:10, "]")]), #group A
                groupB = colMeans(posterior_samples_array[,paste0("b[(Intercept) groupB:", 1:10, "]")]) #group B
)

covb <- list(groupA = cov(posterior_samples_array[,paste0("b[(Intercept) groupA:", 1:10, "]")]),
             groupB = cov(posterior_samples_array[,paste0("b[(Intercept) groupB:", 1:10, "]")]))

sig <- colMeans(posterior_samples_array[,c("Sigma[groupA:(Intercept),(Intercept)]","Sigma[groupB:(Intercept),(Intercept)]")])

EBF(b, covb, sig)

}
\references{
Vieira, F., Zhao, H., & Mulder, J. (2024). To Vary or Not To Vary: A Simple Empirical Bayes Factor for Testing Variance Components. arXiv preprint arXiv:2410.14459.

Vieira, F., Leenders, R., McFarland, D., & Mulder, J. (2024). A Bayesian actor-oriented multilevel relational event model with hypothesis testing procedures. Behaviormetrika, 51(1), 37-74.
}
