% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbinom.bias.R
\docType{data}
\name{nbinom.bias}
\alias{nbinom.bias}
\title{EBF bias for negative binomial tests}
\format{
A vector with the bias for each sample size from 1 to 100.
}
\usage{
nbinom.bias
}
\description{
Bias in the log posterior marginal likelihood for negative binomial tests,
used for calculating empirical Bayes factors.
Bias is computed by the \code{\link{compute.nbinom.bias}} function using Monte Carlo simulation.
Minimum sample sizes (ie fixed number of events) range from 1 to 100.
The uniform prior is used.
}
\details{
The bias depends on the true success probability.
As this is unknown, the bias is maximised over the probability to give
a conservative estimate.
}
\references{
Dudbridge F (2022) A scale of evidence for likelihood ratios.
arXiv:2212.06669

Dudbridge F (in preparation) Empirical Bayes factors.
}
\author{
Frank Dudbridge
}
\keyword{datasets}
